Remove dead code

+++ library/src/appleMain/kotlin/com/tidal/networktime/internal/NtpUdpSocketOperations.kt
@@ -1,10 +1,9 @@
 package com.tidal.networktime.internal
 
 internal actual class NtpUdpSocketOperations {
-
   actual fun prepareSocket(timeoutMilliseconds: Long) {}
 
-  actual fun exchangePacketInPlace(buffer: ByteArray, address: String, portNumber: Byte) {}
+  actual fun exchangePacketInPlace(buffer: ByteArray, address: String, portNumber: Int) {}
 
   actual fun closeSocket() {}
 }
